+++ src/metaWord/src/extension.ts
@@ -1,14 +1,14 @@
 'use strict';
 import * as vscode from 'vscode';
 import { Config } from "./config";
-import { LandingPage } from './landing-page';
+// import { LandingPage } from './landing-page';
 import {MetaSpaceWord} from './space-word';
 
 // this method is called when your extension is activated
 // your extension is activated the very first time the command is executed
 export function activate(context: vscode.ExtensionContext) {
-    const landingPage = new LandingPage(context);
-    landingPage.showIfNeed();
+    // const landingPage = new LandingPage(context);
+    // landingPage.showIfNeed();
     let config = new Config();
     config.loadConfig();
     // Event to update active configuration items when changed without restarting vscode