+++ src/metaWord/src/space-word.ts
@@ -4,7 +4,7 @@ import { Config } from './config';
 enum Mode { Move, Select, Delete }
 export class MetaSpaceWord {
 
-    updateConfig(){
+    updateConfig() {
 
     }
     constructor(context: vscode.ExtensionContext, private config: Config) {
@@ -18,6 +18,17 @@ export class MetaSpaceWord {
             vscode.commands.registerTextEditorCommand('metaGo.cursorSpaceWordLeftDelete',
                 (editor, edit) => this.left(editor, edit, Mode.Delete)
             ),
+
+            vscode.commands.registerTextEditorCommand('metaGo.cursorSpaceWordSpaceLeft',
+                (editor, edit) => this.left(editor, edit, Mode.Move, true)
+            ),
+            vscode.commands.registerTextEditorCommand('metaGo.cursorSpaceWordSpaceLeftSelect',
+                (editor, edit) => this.left(editor, edit, Mode.Select, true)
+            ),
+            vscode.commands.registerTextEditorCommand('metaGo.cursorSpaceWordSpaceLeftDelete',
+                (editor, edit) => this.left(editor, edit, Mode.Delete, true)
+            ),
+
             vscode.commands.registerTextEditorCommand('metaGo.cursorSpaceWordRight',
                 (editor, edit) => this.right(editor, edit)
             ),
@@ -26,11 +37,21 @@ export class MetaSpaceWord {
             ),
             vscode.commands.registerTextEditorCommand('metaGo.cursorSpaceWordRightDelete',
                 (editor, edit) => this.right(editor, edit, Mode.Delete)
+            ),
+
+            vscode.commands.registerTextEditorCommand('metaGo.cursorSpaceWordSpaceRight',
+                (editor, edit) => this.right(editor, edit, Mode.Move, true)
+            ),
+            vscode.commands.registerTextEditorCommand('metaGo.cursorSpaceWordSpaceRightSelect',
+                (editor, edit) => this.right(editor, edit, Mode.Select, true)
+            ),
+            vscode.commands.registerTextEditorCommand('metaGo.cursorSpaceWordSpaceRightDelete',
+                (editor, edit) => this.right(editor, edit, Mode.Delete, true)
             )
         )
     }
 
-    left(editor, edit, mode = Mode.Move, char = ' ') {
+    left(editor, edit, mode = Mode.Move, includeChar = false, chars = [' ', '\t']) {
         let selections: vscode.Selection[] = [];
         for (let s = 0; s < editor.selections.length; s++) {
             const selection = editor.selections[s];
@@ -47,14 +68,31 @@ export class MetaSpaceWord {
                 charIndex = text.length;
 
             let findNoneSpace = false;
+            let findSpace = false;
             for (let j = charIndex - 1; j >= -1; j--) {
-                if (text[j] !== char && text[j] !== undefined) {
+                if (chars.every(c => c !== text[j]) && text[j] !== undefined) {
                     findNoneSpace = true;
-                    continue;
+                    if (!includeChar || !findSpace)
+                        continue;
                 }
-                if (findNoneSpace && (text[j] === ' ' || j === -1) || j === -1) {
+                if (j === -1) {
                     position = new vscode.Position(i, j + 1);
                     break;
+                }
+                if (includeChar) {
+                    if (chars.some(c => c === text[j])) {
+                        findSpace = true;
+                        continue;
+                    }
+                    if (findSpace) {
+                        position = new vscode.Position(i, j + 1);
+                        break;
+                    }
+                } else {
+                    if (findNoneSpace && chars.some(c => c === text[j])) {
+                        position = new vscode.Position(i, j + 1);
+                        break;
+                    }
                 }
             }
             if (i === -1) position = selection.active;
@@ -63,8 +101,7 @@ export class MetaSpaceWord {
 
         editor.selections = selections;
     }
-
-    right(editor: vscode.TextEditor, edit: vscode.TextEditorEdit, mode = Mode.Move, char = ' ') {
+    right(editor: vscode.TextEditor, edit: vscode.TextEditorEdit, mode = Mode.Move, includeChar = false, chars = [' ', '\t']) {
         const lines = editor.document.lineCount;
         let selections: vscode.Selection[] = [];
         for (let s = 0; s < editor.selections.length; s++) {
@@ -81,14 +118,32 @@ export class MetaSpaceWord {
             let position: vscode.Position;
 
             let findNoneSpace = false;
+            let findSpace = false;
             for (let j = charIndex; j <= text.length; j++) {
-                if (text[j] !== char && text[j] !== undefined) {
+                if (chars.every(c => c !== text[j]) && text[j] !== undefined) {
                     findNoneSpace = true;
-                    continue;
+                    if (!includeChar || !findSpace)
+                        continue;
                 }
-                if (findNoneSpace && (text[j] === ' ' || j === text.length) || j === text.length) {
+                if (j === text.length) {
                     position = new vscode.Position(i, j);
                     break;
+                }
+
+                if (includeChar) {
+                    if (chars.some(c => c === text[j])) {
+                        findSpace = true;
+                        continue;
+                    }
+                    if (findSpace) {
+                        position = new vscode.Position(i, j);
+                        break;
+                    }
+                } else {
+                    if (findNoneSpace && chars.some(c => c === text[j])) {
+                        position = new vscode.Position(i, j);
+                        break;
+                    }
                 }
             }